layout CarbDialog alarm and use timer for automation alarms
