+++ app/src/main/java/info/nightscout/androidaps/plugins/general/automation/actions/ActionAlarm.kt
@@ -1,19 +1,20 @@
 package info.nightscout.androidaps.plugins.general.automation.actions
 
 import android.content.Context
-import android.content.Intent
 import android.widget.LinearLayout
 import androidx.annotation.DrawableRes
 import dagger.android.HasAndroidInjector
 import info.nightscout.androidaps.R
-import info.nightscout.androidaps.activities.ErrorHelperActivity
 import info.nightscout.androidaps.data.PumpEnactResult
 import info.nightscout.androidaps.plugins.bus.RxBusWrapper
 import info.nightscout.androidaps.plugins.general.automation.elements.InputString
 import info.nightscout.androidaps.plugins.general.automation.elements.LabelWithElement
 import info.nightscout.androidaps.plugins.general.automation.elements.LayoutBuilder
 import info.nightscout.androidaps.queue.Callback
+import info.nightscout.androidaps.utils.DateUtil
 import info.nightscout.androidaps.utils.JsonHelper
+import info.nightscout.androidaps.utils.T
+import info.nightscout.androidaps.utils.TimerUtil
 import info.nightscout.androidaps.utils.resources.ResourceHelper
 import org.json.JSONObject
 import javax.inject.Inject
@@ -23,26 +24,23 @@ class ActionAlarm(injector: HasAndroidInjector) : Action(injector) {
     @Inject lateinit var resourceHelper: ResourceHelper
     @Inject lateinit var rxBus: RxBusWrapper
     @Inject lateinit var context: Context
+    @Inject lateinit var timerUtil: TimerUtil
 
     var text = InputString(injector)
 
     constructor(injector: HasAndroidInjector, text: String) : this(injector) {
         this.text = InputString(injector, text)
     }
+
     override fun friendlyName(): Int = R.string.alarm
     override fun shortDescription(): String = resourceHelper.gs(R.string.alarm_message, text.value)
     @DrawableRes override fun icon(): Int = R.drawable.ic_access_alarm_24dp
 
-    override fun isValid(): Boolean = text.value.isNotEmpty()
+    override fun isValid(): Boolean = true // empty alarm will show app name
 
     override fun doAction(callback: Callback) {
-        val i = Intent(context, ErrorHelperActivity::class.java)
-        i.putExtra("soundid", R.raw.modern_alarm)
-        i.putExtra("status", text.value)
-        i.putExtra("title", resourceHelper.gs(R.string.alarm))
-        i.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-        context.startActivity(i)
-
+        timerUtil.scheduleReminder(DateUtil.now() + T.secs(10L).msecs(), text.value.takeIf { it.isNotBlank() }
+            ?: resourceHelper.gs(R.string.app_name))
         callback.result(PumpEnactResult(injector).success(true).comment(R.string.ok))?.run()
     }
