+++ app/src/main/java/info/nightscout/androidaps/utils/TimerUtil.kt
@@ -1,0 +1,25 @@
+package info.nightscout.androidaps.utils
+
+import android.content.Context
+import android.content.Intent
+import android.provider.AlarmClock
+import info.nightscout.androidaps.R
+import info.nightscout.androidaps.utils.resources.ResourceHelper
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class TimerUtil @Inject constructor(
+    private val context: Context,
+    private val resourceHelper: ResourceHelper,
+) {
+
+    fun scheduleReminder(time: Long, text: String? = null) = Intent(AlarmClock.ACTION_SET_TIMER).apply {
+        val length: Int = ((time - DateUtil.now()) / 1000).toInt()
+        flags = flags or Intent.FLAG_ACTIVITY_NEW_TASK
+        putExtra(AlarmClock.EXTRA_LENGTH, length)
+        putExtra(AlarmClock.EXTRA_SKIP_UI, true)
+        putExtra(AlarmClock.EXTRA_MESSAGE, text ?: resourceHelper.gs(R.string.app_name))
+        context.startActivity(this)
+    }
+}
